♻️ Trekk ut felles funksjoner for journalføring i egen fil
